Extract BuildResults into its own component
